requirements.txt
pypdf>=4.0.0
Pillow>=10.0.0
Jinja2>=3.1.0
Flask>=3.0.0
reportlab>=4.2.0
